Move elevation data fetching out of the ground impl
